Rewrote shooter to allow for angling of tragectory
Left bumper raises, right lowers, start button on controller toggles it on and off.
This is meant to resolve #14, but does not exactly follow the method I described in that issue.
Basically, I just put everything in the Shooter class, periodically checking if the buttons I wanted had been clicked using the XboxController.get____Pressed() and then set the speed of the motors depending on whether they should be running or not
Also commented out the start Command in the Shooter class, as it didn't seem necessary since the class handles it's own stuff in the periodic